Ejercicio5/requirements.txt
azure-data-tables==12.4.4
python-dotenv